adds game.cpp as method implementation file, changes CMakeLists.txt and recompiles build files
