+++ src/game.cpp
@@ -1,0 +1,12 @@
+#include <iostream>
+#include "volsweeper.hpp"
+using namespace std;
+
+int main(int argc, char* argv[]) {
+    if (argc != 3) {
+        cerr << "usage: ./volsweeper [FLAG] [SIZE]\n";
+        return 1;
+    }
+
+    return 0;
+}
